+++ tests/core/test_doe_scenario.py
@@ -22,8 +22,11 @@
 from __future__ import unicode_literals
 
 import pytest
+from numpy import array
 
+from gemseo.algos.design_space import DesignSpace
 from gemseo.api import create_discipline, create_scenario
+from gemseo.core.analytic_discipline import AnalyticDiscipline
 from gemseo.core.doe_scenario import DOEScenario
 from gemseo.problems.sobieski.core import SobieskiProblem
 from gemseo.problems.sobieski.wrappers import (
@@ -93,7 +96,7 @@ def mdf_variable_grammar_doe_scenario(request):
 
 @pytest.mark.usefixtures("tmp_wd")
 @pytest.mark.skip_under_windows
-def test_parallel_doe_hdf_cache():
+def test_parallel_doe_hdf_cache(caplog):
     disciplines = create_discipline(
         [
             "SobieskiStructure",
@@ -133,6 +136,8 @@ def test_parallel_doe_hdf_cache():
         "algo_options": {"n_processes": 2, "n_samples": n_samples},
     }
     scenario.execute(input_data)
+    expected_log = "Double definition of algorithm option n_samples, keeping value: {}."
+    assert expected_log.format(n_samples) in caplog.text
 
 
 @pytest.mark.parametrize(
@@ -159,3 +164,20 @@ def test_doe_scenario(mdf_variable_grammar_doe_scenario):
         len(mdf_variable_grammar_doe_scenario.formulation.opt_problem.database)
         == n_samples
     )
+
+
+def test_warning_when_missing_option(caplog):
+    """Check that a warning is correctly logged when an option is unknown."""
+    discipline = AnalyticDiscipline(name="func", expressions_dict={"y": "2*x"})
+    design_space = DesignSpace()
+    design_space.add_variable("x", l_b=0.0, u_b=1.0)
+    scenario = DOEScenario([discipline], "DisciplinaryOpt", "y", design_space)
+    scenario.execute(
+        {
+            "algo": "CustomDOE",
+            "algo_options": {"samples": array([[1.0]]), "unknown_option": 1},
+        }
+    )
+    expected_log = "Driver CustomDOE has no option {}, option is ignored."
+    assert expected_log.format("n_samples") not in caplog.text
+    assert expected_log.format("unknown_option") in caplog.text